+++ src/Rules/Accessibility/MetaViewportPresent.php
@@ -13,9 +13,13 @@ class MetaViewportPresent extends BaseRule
 {
     public function __construct()
     {
-        $this->name = "metaViewportPresent";
         $this->value = 14;
         $this->type = RuleType::Accessibility;
+    }
+
+    public function name(): string
+    {
+        return "metaViewportPresent";
     }
 
     protected function passes(): bool