+++ src/Rules/Accessibility/UseLandmarkTags.php
@@ -10,9 +10,13 @@ class UseLandmarkTags extends BaseRule
 {
     public function __construct()
     {
-        $this->name = "useLandmarkTags";
         $this->value = 14;
         $this->type = RuleType::Accessibility;
+    }
+
+    public function name(): string
+    {
+        return "useLandmarkTags";
     }
 
     protected function passes(): bool